update: series line
